requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
boto3==1.14.6
botocore==1.17.6
certifi==2020.4.5.2
chardet==3.0.4
click==7.1.2
discord.py==1.3.3
docutils==0.15.2
Flask==1.1.2
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
MarkupSafe==1.1.1
multidict==4.7.6
python-dateutil==2.8.1
python-dotenv==0.13.0
requests==2.24.0
s3transfer==0.3.3
six==1.15.0
urllib3==1.25.9
websockets==8.1
Werkzeug==1.0.1
yarl==1.4.2
